msm: audio: qdsp6v2: Fix voice RTAC
There was data type mismatch which caused the corruption during
memory copy. It leads to the voice RTAC failure.

CRs-Fixed: 391392
(cherry picked from commit aca95463302f334a6fe752201812c40019fe53c1)
(cherry picked from commit fd155173f9f3ce98ebd7674d560656168d0af49c)

+++ arch/arm/mach-msm/qdsp6v2/rtac.c
@@ -752,7 +752,7 @@ u32 send_voice_apr(u32 mode, void *buf, u32 opcode)
 	u32	count = 0;
 	u32	bytes_returned = 0;
 	u32	payload_size;
-	u16	dest_port;
+	u32	dest_port;
 	struct	apr_hdr	voice_params;
 	pr_debug("%s\n", __func__);
